Create mutable document view
