+++ lib/nanoc/base/views/mutable_item.rb
@@ -1,22 +1,5 @@
 module Nanoc
   class MutableItemView < Nanoc::ItemView
-    # Sets the value for the given attribute.
-    #
-    # @param [Symbol] key
-    #
-    # @see Hash#[]=
-    def []=(key, value)
-      unwrap.attributes[key] = value
-    end
-
-    # Updates the attributes based on the given hash.
-    #
-    # @param [Hash] hash
-    #
-    # @return [self]
-    def update_attributes(hash)
-      hash.each { |k, v| unwrap.attributes[k] = v }
-      self
-    end
+    include Nanoc::MutableDocumentViewMixin
   end
 end
