+++ lib/nanoc/base/views/mutable_layout.rb
@@ -1,12 +1,5 @@
 module Nanoc
   class MutableLayoutView < Nanoc::LayoutView
-    # Sets the value for the given attribute.
-    #
-    # @param [Symbol] key
-    #
-    # @see Hash#[]=
-    def []=(key, value)
-      unwrap.attributes[key] = value
-    end
+    include Nanoc::MutableDocumentViewMixin
   end
 end
