+++ spec/nanoc/base/views/mutable_item_spec.rb
@@ -1,28 +1,4 @@
 describe Nanoc::MutableItemView do
-  describe '#[]=' do
-    let(:item) { Nanoc::Int::Item.new('content', {}, '/asdf/') }
-    let(:view) { described_class.new(item) }
-
-    it 'sets attributes' do
-      view[:title] = 'Donkey'
-      expect(view[:title]).to eq('Donkey')
-    end
-  end
-
-  describe '#update_attributes' do
-    let(:item) { Nanoc::Int::Item.new('content', {}, '/asdf/') }
-    let(:view) { described_class.new(item) }
-
-    let(:update) { { friend: 'Giraffe' } }
-
-    subject { view.update_attributes(update) }
-
-    it 'sets attributes' do
-      expect { subject }.to change { view[:friend] }.from(nil).to('Giraffe')
-    end
-
-    it 'returns self' do
-      expect(subject).to equal(view)
-    end
-  end
+  let(:entity_class) { Nanoc::Int::Item }
+  it_behaves_like 'a mutable document view'
 end
