+++ spec/nanoc/base/views/mutable_layout_spec.rb
@@ -1,11 +1,4 @@
 describe Nanoc::MutableLayoutView do
-  describe '#[]=' do
-    let(:layout) { Nanoc::Int::Layout.new('content', {}, '/asdf/') }
-    let(:view) { described_class.new(layout) }
-
-    it 'sets attributes' do
-      view[:title] = 'Donkey'
-      expect(view[:title]).to eq('Donkey')
-    end
-  end
+  let(:entity_class) { Nanoc::Int::Layout }
+  it_behaves_like 'a mutable document view'
 end
